I2C MASTER has been added continue
The I2C SLAVE functionality now can be verified
by the in-chip implemented I2C MASTER functionality.

The I2C MASTER SDA is on RC5 pin, SCL is on RC4 pin.
The I2C SLAVE SDA is on RC3 pin, SCL is on RC2 pin.

The SDA pins have to be connected with a wire,
SCL pins also have to be connected with a wire.

In the MAIN loop (main.c) there is a section
inside the "if (isI2CMasterDeviceUsed) {}"
IP-statement.

Here are 4 examples for read/write operations.